fix: move enter room logic in userDashboard
